Refactored some debug code
