+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Debug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db88e3163cfc0c1489e48fea43109f3e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
