+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Debug/nickmaltbie.OpenKCC.Debug.asmdef
@@ -1,0 +1,22 @@
+{
+    "name": "nickmaltbie.OpenKCC.Debug",
+    "rootNamespace": "nickmaltbie.OpenKCC.Debug",
+    "references": [
+        "Unity.InputSystem",
+        "Unity.TextMeshPro",
+        "nickmaltbie.ScreenManager",
+        "nickmaltbie.ScreenManager.Actions",
+        "nickmaltbie.TestUtilsUnity",
+        "nickmaltbie.StateMachineUnity",
+        "nickmaltbie.OpenKCC"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
