+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Debug/nickmaltbie.OpenKCC.Debug.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 384eb3392edcdb24aae6c61802d65573
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
